requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.3
frozenlist==1.3.3
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
meraki==1.27.0
multidict==6.0.3
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.45
urllib3==1.26.11
Werkzeug==2.2.1
yarl==1.8.2
zipp==3.8.1
